% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{etl_one_file}
\alias{etl_one_file}
\title{Run the entire ETL process for one file index}
\usage{
etl_one_file(idx)
}
\arguments{
\item{idx}{}
}
\value{

}
\description{
Assumes files have format /././file_<idx>.json
}
