% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_file_idx}
\alias{extract_file_idx}
\title{Extracts integers from filenames}
\usage{
extract_file_idx(
  dir = "/home/rstudio/data/data-challenge/data/interconnect_json/"
)
}
\arguments{
\item{dir}{directory where files are stored on nfs}
}
\value{

}
\description{
Extracts integers from filenames
}
