% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{glob_parquets}
\alias{glob_parquets}
\title{Read in multipart parquet
https://stackoverflow.com/a/58450079}
\usage{
glob_parquets(file_list)
}
\arguments{
\item{file_list}{}
}
\value{

}
\description{
Read in multipart parquet
https://stackoverflow.com/a/58450079
}
