% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_to_sqlite}
\alias{load_to_sqlite}
\title{Loads the three data sets into a sqlite database}
\usage{
load_to_sqlite(data, db_path = here::here("data.db"), idx = NULL)
}
\arguments{
\item{data}{List of dataframes}

\item{db_path}{database path}
}
\value{

}
\description{
Loads the three data sets into a sqlite database
}
