% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_and_clean_data}
\alias{parse_and_clean_data}
\title{Parse and Clean Data}
\usage{
parse_and_clean_data(df)
}
\arguments{
\item{df}{}
}
\value{

}
\description{
The raw data is a nested dataframe. The command "unnest"
will take the arrays and flatten them, repeating each
row of the parent.
}
\details{
I create three dataframes here, to make it easy to
save them out to SQL. I retain the file_id in the bonds table
and the bond_id in the sample table to make it easy to join.
}
