% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prep_group_id_data}
\alias{prep_group_id_data}
\title{Split data by bond_group_id}
\usage{
prep_group_id_data(data)
}
\arguments{
\item{data}{list with a dataframe called bonds_with_samples inside}
}
\value{

}
\description{
splits bond sample data into partitions by
bond_group_id
}
