% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_as_parquet}
\alias{save_as_parquet}
\title{Write as parquet}
\usage{
save_as_parquet(data, dir_root, idx)
}
\arguments{
\item{idx}{}
}
\value{

}
\description{
Creates individual tables for each section of json object
}
