% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_column_if_doesnt_exist}
\alias{add_column_if_doesnt_exist}
\title{Checks if every column in the dataset exists in the
SQLITE table. If it doesn't, it adds one.}
\usage{
add_column_if_doesnt_exist(df, pool, tab)
}
\arguments{
\item{pool}{Connection}

\item{data}{List of dataframes}

\item{table}{}
}
\value{

}
\description{
Checks if every column in the dataset exists in the
SQLITE table. If it doesn't, it adds one.
}
